+++ R/reports.R
@@ -53,11 +53,11 @@ write_summary_charts = function(worksheet_path, output_path) {
                     params=list(data_path=worksheet_path))
 }
 
-#' Create an unmixing error report for simplex samples
+#' Create an unmixing quality report for simplex samples
 #'
 #' `export_path` should be the path to
 #' an export folder containing component data files for singleplex samples.
-#' An unmixing error report is created for the samples and saved in
+#' An unmixing quality report is created for the samples and saved in
 #' the source directory.
 #'
 #' The report generator tries to identify the Opal fluor for each source file
@@ -66,13 +66,13 @@ write_summary_charts = function(worksheet_path, output_path) {
 #' It also recognizes three leading digits as the number of an Opal fluor.
 #' @param export_path Path to a directory containing component_data files.
 #' @export
-unmixing_error_report = function(export_path=NULL) {
+unmixing_quality_report = function(export_path=NULL) {
   stopifnot(dir.exists(export_path))
 
-  rmd_path = system.file("rmd", "Unmixing_error.Rmd",
+  rmd_path = system.file("rmd", "Unmixing_quality_report.Rmd",
                          package="phenoptrReports")
 
-  output_path = file.path(export_path, 'Unmixing_error.html')
+  output_path = file.path(export_path, 'Unmixing_quality_report.html')
   rmarkdown::render(rmd_path, output_file=output_path,
                     params=list(export_path = export_path))
 }